Fix extracting requires for extras
Consider these cases:

  langchain
  langchain[llms]
  langchain[llms,cli,embeddings]

Each requirement listed in langchain's Requires-Dist needs to be evaluated
against an environment that can only contain a single 'extra' key.

So for the first case, we pass an empty environment and the library fills
in "" as the default for 'extra'.

For the second case, we evaluate once against an environment with 'extra'
set to 'llms'.

For the final case, we need to evaluate repeatedly, once each with an
environment with 'extra' set to 'llms', 'cli', and 'embeddings'.