remove unnecessary patches adding ninja to build requirements
In the course of investigating why these are not needed upstream, I
found that removing them no longer breaks the build. I suspect we had
a sequencing issue before, but it's not clear. In any case,
meson_python includes its own build backend which adds ninja to the
list of build requirements dynamically if the ninja command is not
found in the $PATH when the dependencies are evaluated [1]. That
dependency is propagated to numpy, so that patch is not needed either.

[1] https://github.com/mesonbuild/meson-python/blob/main/mesonpy/__init__.py#L1031-L1032

Fixes #45
Fixes #46