ninja: hack to make our wheel wrap the ninja-build RPM
During the build of this wheel, a source archive was being downloaded
from github.

It turns out the ninja wheel is a very simple Python package that wraps
a ninja-build binary. We can just use the system ninja-build RPM.

The Python package also includes the ninja_syntax module, which it copies
from the source archive it downloads. For now, just include that in our
workaround patch.