fix override lookup for dists with '.' in the name
Some dists that install into namespaced packages are released or
included in dependency lists with the '.' in the name. For example,
devpi has a dependency on "repoze.lru". That name is a valid python
module name, but we don't use nested namespace packages for overrides,
so convert the '.' to '_' to make a valid module name without a
package hierarchy.

Fixes #116