use unique venvs per script
Ensure that the mirror and build scripts do not share a virtualenv so
that packages installed by the mirror tool are not present when trying
to run a build. This isn't complete isolation, but does let us test
that the builder reports errors properly when some dependency is not met.