add more logic trying to discover downloaded source with nonstandard name
Include the requirement name, as given, as a way for the caller to
influence where the system looks for the downloaded source archive and
the unpacked source code. This handles cases like oslo.messaging,
where there is a '.' in both names that would normally be removed by
the name transformation logic.