Fix torch.download_source()
commit 9e7fb86ae broke this by adding a new parameter not reflected
in the override.

First, chain to sources.default_download_source() when we're downloading
the cached copy - that should work as normal.

Second, use sources.resolve_sdist() with a new only_sdists parameter
rather than copying the implementation. This also makes it easier to
intercept the resolution during unit testing.

Finally, add a unit test so this doesn't break in future.