Make sdist-server-url a required argument for download
Currently, if this arg isn't supplied, download_source() doesn't
actually fall back to PyPI as a default. This is probably a good
thing because we don't want o accidentally fallback to PyPI because
of a bug.

There are two places we are deliberately using PyPI:

- Downloading source during bootstrap
- Resolving dependencies when one is missing

In both cases, let's have the calling code explicitly specify
PyPI as the server.