resolver: use DEBUG_RESOLVER environment variable to control debug messages
We want to turn these on and off a lot to control how much output we
have to deal with in the log file vs. how much information we need to
understand a failure. Use an environment variable to do that, since
we're already using debug log levels for more useful messages.