improve debugging when plugins do not load
Die if we find a plugin but can't import it. This ensures we don't
silently ignore errors in the plugin modules.

Improve debug message when we just don't find a plugin, showing which
ones we do find.

Log when we do find an override so it's clear which path we're taking.