feature: log the output of the build command
When we run the step to build the wheel, tell pip to write a log file
in the directory above where the source is located. Using this
location ensures the file is preserved when bootstrap cleans up the
source tree, and makes it easy to collect all of the files if we are
building a sequence.