make e2e jobs more robust
Some of the e2e jobs had hard-coded expectations for version numbers
of dependencies of packages involved in the tests. Those versions will
change over time as new releases are created, so this commit reworks
the tests to be more flexible about the version numbers and removes
the dependencies that have been dropped from the test packages.