feature: let overrides do version resolution
Some overrides are for doing things like building from a fork of a
project with fixes that aren't upstream. In those cases, the builder
might have to determine the version in some way other than looking at
pypi.org. Provide a hook for that, and include a sample implementation
that uses tags on github.com.