evaluate marker expressions for toplevel dependencies
If we are given a requirements file as the set of toplevel dependencies, it is reasonable to expect some entries to have marker expressions controlling when the dependency is relevant. This change ensures those expressions are evaluated so that dependencies for different versions of python or other platforms are ignored.