fix reading requirements files for bootstrapping
The requirements files are always passed from click as a sequence, even
if there is only one value. We need to iterate over them and convert
the filename to a string and parse each file individually. This change
fixes the parsing function and updates the unit test to ensure that it
works correctly by actually trying to parse files instead of mocking the
return value.