fix filtering based on markers during bootstrapping
If we're given a requirements file as input, we
might be iterating over a list of requirements
with marker expressions that limit their use to
specific platforms or python versions. Evaluate
the markers to filter out anything we shouldn't
build. Only apply the filter to toplevel
requirements (items without a why list leading up
to them) because other dependencies are already
filtered based on their markers in the context of
their parent, so they include values like the
parent's extras settings.

Fixes #182