build_sdist hook with extra_environ and version
The `build_sdist` hook and `pep517_build_sdist` function now receive
`extra_environ` parameter. This allows the caller to modify the
environment for building an sdist.

The `build_wheel` and `build_sdist` hooks also receive a `version`
parameter.

Both changes are needed for PyTorch. The packages needs extra env
variables to be set to to the current version and build number.

Breaking change: any existing `build_wheel` and `build_sdist` hooks
must be updated with the new arguments.

Signed-off-by: Christian Heimes <cheimes@redhat.com>