Only vendor Rust when project uses setuptools-rust / maturin
`vendor_rust` now inspects `pyproject.toml` and looks for
`setuptools-rust` and `maturin` in `[build-system]requires`. When
`pyproject.toml` does not exist or does not list a Rust build plugin as
requirement, then Rust crates are not vendored.

The change indirectly addresses the issue with Meson's broken Rust
crates in `test cases/rust`.

Signed-off-by: Christian Heimes <cheimes@redhat.com>