only write installation requirements to constraints
Keep a complete graph of all of the dependencies and their reasons so we
can use that to find only installation dependencies of the root items to
include in the constraints file.

Remove tests that required the generated constraints file to match the
build-order file, since that no longer applies.

Remove the command to convert a build-order file
to a constraints list. That command produces
useless output. We will add a set of commands for
manipulating the graph, later.