Fix progress bar recursion in bootstrap command
The bootstrap command uses a recursive algorithm to fetch sources and to
build sdists + wheels. The complete set of dependencies is unknown until
all packages have been downloaded and built.

The bootstrap command now passes an optional `progressbar` object
around. Each handler updates the total amount of items in the beginning
of a loop, and notifies the progress bar when an item has finished.

Signed-off-by: Christian Heimes <cheimes@redhat.com>