remove "why" information from build-order files
We added a dependency chain field to the
build-order file so we could easily tell why
something was included in the build. The
information was limited to the first parent that
depended on each item. Since then, we have added a
more complete graph file, which includes all
dependencies. Removing the "why" field from the
build-order file will eliminate a redundant piece
of info and reduce the size of differences in
those files when they are tracked by version
control tools, making the differences easier to
understand.