Pydantic model for global and package settings
Introduce Pydantic to validate and manage `settings.yaml` and
package-specific `package.yaml` config files. Pydantic handles
validation and default management of settings files.

The package settings files now contains environment variables overrides
for package and package variants as well as pre-built flag for each
variant. The `env` files and most fields in from the old `settings.yaml`
will be replaced eventually.

A new command `fromager migrate-config` handles migration of config
and env files from the old format to the new format.

The new code is not used, yet. Code will be switched to the new API in a
subsequent PR.

Signed-off-by: Christian Heimes <cheimes@redhat.com>