Support variant specific patches
When building wheels for different variants, the code base might be
different from one variant to the other, for example when the code comes
from a fork with variant specific features. And with a different code
base, the patches may fail to be applied.

This changes proposes to support patch files stored in a variant
specific path under the existing unversioned and versioned patch
directories. Below is an example of a patch tree:

```
.../patches
.../patches/deepspeed
.../pacthes/deepspeed/0001-unversioned.patch
.../pacthes/deepspeed/0002-unversioned.patch
.../patches/deepspeed/brie/0001-unversioned.patch
.../patches/deepspeed/feta/0002-unversioned.patch
.../patches/deepspeed-0.5.0/0001-versioned.patch
.../patches/deepspeed-0.5.0/brie/0003-versioned.patch
```

When applying the patches for the `brie` variant, the following patches
would be applied:

```
.../pacthes/deepspeed/0001-unversioned.patch
.../patches/deepspeed/brie/0001-unversioned.patch
.../pacthes/deepspeed/0002-unversioned.patch
.../patches/deepspeed-0.5.0/0001-versioned.patch
.../patches/deepspeed-0.5.0/brie/0003-versioned.patch
```

When applying the patches for the `feta` variant, the following patches
would be applied:

```
.../pacthes/deepspeed/0001-unversioned.patch
.../pacthes/deepspeed/0002-unversioned.patch
.../patches/deepspeed/feta/0002-unversioned.patch
.../patches/deepspeed-0.5.0/0001-versioned.patch
```

Signed-off-by: Fabien Dupont <fdupont@redhat.com>