ensure packages with common prefixes do not mix patches
Fix a problem introduced in #488 that could cause packages with a prefix
common to another package's full name to have their patches included
when scanning for patches of the package with the shorter name. For
example, with patches for both 'foo' and 'foo-bar' when patches for
'foo' are loaded they would have included patches for 'foo-bar'.