bin/bootstrap: Add -k option to keep container running
It is occasionally convenient to keep the build container running when
debugging a wheel build failure.  Add a -k, or "keep alive", parameter to
the script that takes a number of seconds to keep the container running as
an argument.

ex) ./bin/bootstrap.sh -k 86400 Containerfile.rocm-ubi9 aotriton==0.7b aotriton==0.7b

Signed-off-by: Prarit Bhargava <prarit@redhat.com>