use local cache when bootstrapping
We added an option to check an external cache for wheels matching the
version and build tag computed during bootstrapping, but the
implementation of that feature lost the ability to skip rebuilding
something that was already on the local disk. This change restores that
behavior, which is especially useful for developers bootstrapping
locally.