Refactor constraints module
Refactor and modify the API of `fromager.constraints` module to prepare
the code for multiple constraints files.

The `Constraints` class no longer takes an mapping of package name to
`Requirement`. Instead it has a new method `add_constraint()` that takes
a requirements strings. The method does not take a `Requirements` object,
because we might want to modify and merge constraints in the future.

The `load()` module-level function has been replaced by
`Constraints.load_constraints_file()` method.

Signed-off-by: Christian Heimes <cheimes@redhat.com>