resolver.py: Fix is_satisfied_by()
is_satisfied_by() returns false if a requirement is defined without a
specifier.  For example using 'aotriton' (no other arguments) results in a
failure regardless of any constraint specified.  In this case it is safe
to return True and return a tested value for the constraint.

Fix is_satisfied_by() to handle requirements without specifiers.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>