support bootstrapping from git checkouts of projects
Support git+https URLs in the input requirements list.

"Resolve" the version by cloning the repo at the specified reference and
generating the metadata to extract the version string.

Rely on the git clone instead of downloading a source distribution or
allowing an override to specify where the source should come from.

Rely on the git clone with preparing source, and ignore any overrides.

Force sdists to be generated using the PEP-517 approach.

Add e2e tests for cases with and without tags in the URLs.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>