only allow source URLs for top-level packages
Use the requirement type for a package to determine if we should resolve
the source for that package when there is a URL in the requirement
specifier. This ensures that we do not pull arbitrary dependencies from
git sources, while still allowing us to explicitly pull a package from a
git tree for nightly builds or similar operations.