add hook entry point for bootstrap
Add a hook entry point invocation for after we have bootstrapped a
package. It is given the source distribution filename and the wheel
filename, just like for the post-build hook, except that the sdist can
be None if the wheel is pre-built and the wheel can be None if we are
bootstrapping in sdist-only mode.