let bootstrapper reuse a wheel on the local filesystem
Update the bootstrapper to reuse wheels already present on the local
filesystem in the work directory instead of only going to the cache
server. This ensures that someone working on multiple related wheels can
reuse one expensive one when building the next.

Update the test for caching in the bootstrap logic to focus on the cases
of an existing wheel on the filesystem and a wheel available from any
cache server. This removes the separate test for using a local cache
server versus using a remote one, since there should effectively be no
difference in those cases.