Log variant name in a few places
Include the variant name in log lines of a few steps like sdist and
wheel builds. This is useful when running multiple builds in parallel.
In downstream we currently rely on the fact that builds have the variant
name in some paths. This causes problems with caching and will go away.

Signed-off-by: Christian Heimes <cheimes@redhat.com>