do not use bundled setuptools in virtualenvs
Change the flag that causes us to use a bundled version of setuptools
in the build virtualenv so no setuptools is included by default.
The bundled version is too old to build some packages, especially
in the face of recent breaking changes that have been worked around
in things that use setuptools.

Instead of using the bundled version, rely on the fact that in
src/fromager/dependencies.py we set the default build backend
requirements to include setuptools if there is no valid information in
the pyproject.toml. This should ensure that we detect setuptools as a
dependency of projects that are likely to use it because it is the
default backend.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>