fix: ensure constraints.txt production considers toplevel pins
Fix an issue with the `graph to-constraints` command and constraints.txt
generation in `bootstrap` where we were ignoring pinned versions in
top level dependencies if that same package was also a dependency of
something else. If we have conflicting versions in that case there is no
way to resolve them and we should report an error.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>