fix: move version resolution to method in bootstrapper
Move the logic for resolving a version of a single package into a
method in the bootstrapper so it can be reused in the bootstrap
command for toplevel dependencies, instead of having 2 implementations.
Remove the restriction in bootstrapper that toplevel dependencies be
resolved before bootstrapping starts. We still do that, but we do not
enforce it in the bootstrapper because we are now using the bootstrapper
code to do that resolution.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>