do not log errors from inside the resolver
Do not log errors when we have trouble accessing a page in the
package index inside the resolver. The caller may interpret this
as not an error, especially during bootstrapping when it means we
probably just have not yet built the package. Let the caller decide
how to handle the error.