fix: make build-sequence work with requirements with git URLs
Ensure bootstrap writes out info indicating that the requirement needs
to come from the git URL.

Refactor some of the git clone handling from the bootstrapper so it
can be reused.

If download_source finds a git URL in a requirement and the directory
with the source does not already exist, have it clone it. That
ensures that the case where we need to clone the repo to determine
the version (during bootstrap) and the case where we know the version
but do not have the source (during build-sequence) both work.

Add an e2e test case.