fix: lower log level of message when we cannot download a file
It is not necessarily an error to fail to download a file when resolving
packages, because sometimes we try to download the file to see if the
package exists in the index. Log the failure as debug instead of error,
and raise the error to let the caller decide if it should be considered
more important or ignored.