Adding documentation around building the project
Fixes #652

Modified the pyproject.toml as it was causing teh hatch build command to fail
as multiple python versions were part of the default environment.

Also the matrix belongs on specific environments (like test) where multi-version
behavior is actually needed, not on the default environment which should be
simple and predictable.

Signed-off-by: Lalatendu Mohanty <lmohanty@redhat.com>