Add matcher argument to resolver providers
The `GenericProvider`, `GitHubTagProvider`, and `GitLabTagProvider` now
accept a matcher argument. The argument can be a callback or a re
pattern object. The new feature enables more powerful matching and
transformation of tags into Version objects.

`GitLabTagProvider.tag_regex` is no longer supported. The new feature is
a more general implementation of the previous feature.

Signed-off-by: Christian Heimes <cheimes@redhat.com>