feat: Add helper to git clone and tarball
The new helper `git_clone_and_tarball()` is an opinionated function to
clone a git repo from tag/ref with submodules and add the content into
a tar ball. The tar ball does not include a `.git` directory, but it
comes with a `.git_archival.txt` for `setuptools-scm`. An optional
callback can be used to rewrite clone URLs.

The base name of the tar ball and first-level directory are always based
on the `prefix` argument.

Signed-off-by: Christian Heimes <cheimes@redhat.com>