fix: use unique file name for build log
The build command was using `{package}_current.log` to store the build log
for *package*. This clashes when parallel build was building two
versions of a package at the same time. In my test case, it was building
two versions of setuptools in parallel.

The command now uses `{name}-{version}.log`.

Signed-off-by: Christian Heimes <cheimes@redhat.com>