fix: improve logging in build-parallel
Improve logging in the build-parallel command to make it clear why some
packages can start building and others cannot.

Sort the nodes to be built so that the loop processes them in a
deterministic order. This should not affect the actual build order, but
it makes it easier to compare log output and to find packages in the
output because the names are sorted.

Fix up some type annotations that were wrong or confused the linter.