fix: correctly handle existing wheel files in build-parallel
Instead of trying to filter out the packages for which we already have
wheels, just treat that as a case for the build thread to deal
with. This simplifies the logic for processing the nodes and ensures
the checks are performed at a point in time where the information is
needed and can be acted on properly.