fix: filter per-package log file based on build thread
Add a ThreadLogFilter class that filters log messages to only include
the ones that come from a specified thread.

Use ThreadLogFilter to ensure that the per-package build logs created
by build-parallel do not mix content from multiple packages.

Change the naming scheme for the per-package build logs to be the
package name and version, without the rest of the wheel filename
details.