fix: correct handling of existing wheels in build commands
The _is_wheel_built function was returning a boolean to indicate to the
caller whether the filename returned should be used. Change the function
to return a Path or None, so the caller only has to check the one
return value.

Return the full path to where we downloaded the wheel, or return None if
we do not download one. Add more logging to make it easier to follow
what is happening, especially when the candidate is not usable.

Log more information about what we are doing when we decide to download
a wheel vs. recognize that the wheel is present on the local server.

Update the e2e tests that look for specific log messages to match the
new patterns.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>