fix: fix install dependency tracking
The bootstrapper was changing the dependency type for installation
dependencies of build dependencies. That led to invalid graphs, showing
for example that torch had no installation dependencies because it was
a build dependency of vllm.

Fix that error, and improve logging to show the dependencies discovered
and where processing is starting.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>