fix: use graphlib in build-parallel core
Our custom dependency tracker for build-parallel code has a bug. It does
not track and handle installation requirements of build requirements
correctly. This causes build bugs when the installation requirement tree
of a build dependency is deep.

This PR completely rewrites and replaces the core logic of
build-parallel command with stdlib's `graphlib` module. The
`TopologicalSorter` detects cycles and tracks when a new node becomes
ready. Nodes with a dependency on an exclusive build node now become
built-ready earlier.

I also changed the code to use a single instance of
`ThreadPoolExecutor`. It's a tiny bit more efficient and saves one level
of indention.

See: #755
Signed-off-by: Christian Heimes <cheimes@redhat.com>