fix: build installation dependencies before building consumers
Change the way we determine whether a package is ready to be built in
build-parallel so that we build its installation dependencies before we
build it.

Logically it should not be possible for package A installation
dependencies to depend in any way on A, so this ordering should be
safe. The change ensures that when we mark A as built, it is actually
ready to be consumed in build environments for other packages.

Addresses #755

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>