fix: fix build-parallel to include installation dependencies
build-parallel needs to build the installation dependencies of a package
before declaring that package as ready. This change uses graphlib to
build a toplogical sorter for the dependency graph and then works
through the iterations of ready nodes. For each set of ready nodes, any
exclusive nodes are built one at a time and then all of the remaining
ready nodes are built together in parallel.

Fixes #755

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>