chore: make Candidate a dataclass
Convert the `Candidate` class to an immutable dataclass.

Instead of an URL to the metadata file, a candidate object now only
holds an `has_metadata` flag. The metadata URL is always
`self.url + 'metadata'`.

`extras` is now a sorted tuple instead of `Iterable[str] | None`. This
simplifies the logic, too.

Signed-off-by: Christian Heimes <cheimes@redhat.com>